% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfit_function.R
\name{logfit}
\alias{logfit}
\title{A function to get some common pseudo R2 measures from a glm}
\usage{
logfit(model)
}
\arguments{
\item{model}{The fitted glm object}
}
\description{
This function takes a fitted glm object and returns some common psuedo R2 measures.
}
\examples{

# A fictitious example predicting vote for party "CDU" (0: no, 1: yes) 
# with political orientation (1: left, ..., 10: right)
lr <-  c(1, 1, 2, 3, 4, 5, 5, 5, 5, 6, 6, 6, 6, 7, 8, 8, 9, 10)
cdu <- c(0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1)
cdu <- factor(cdu, levels = c(0, 1), labels = c("no", "yes"))
df <- data.frame(lr, cdu)
lm1 <- glm(cdu ~ lr, data = df, family = binomial(link = "logit"))
logfit(lm1)
}
\keyword{glm}
